+++ lib/xmat-snack-bar/index.ts
@@ -1,18 +1,21 @@
 import {NgModule} from "@angular/core";
 import {CommonModule} from "@angular/common";
 import {XmatSnackBarComponent} from "./ts/xmat-snack-bar.component";
+import {MatSnackBarModule} from "@angular/material";
 
 export * from "./ts/xmat-snack-bar.component";
 
 @NgModule({
     imports: [
-        CommonModule
+        CommonModule,
+        MatSnackBarModule
     ],
     declarations: [
         XmatSnackBarComponent
     ],
     exports: [
-        XmatSnackBarComponent
+        XmatSnackBarComponent,
+        MatSnackBarModule
     ],
     entryComponents: [
         XmatSnackBarComponent